requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-huggingface
faster-whisper
torch
torchvision
opencv-python-headless
pillow
python-dotenv
faiss-cpu
